+++ tests/acceptance/dashboard-test.ts
@@ -12,7 +12,8 @@ module('Acceptance | dashboard', hooks => {
         // A fully loaded dashboard should have no major troubles
         const currentUser = server.create('user');
         const nodes = server.createList('node', 10, {}, 'withContributors');
-        server.loadFixtures('nodes');
+        server.create('node', { id: 'z3sg2', linkedNodeIds: ['1', '2', '3', '4', '5'], title: 'NNW' });
+        server.create('node', { id: '57tnq', linkedNodeIds: ['6', '7', '8', '9', '10'], title: 'Popular' });
         for (let i = 4; i < 10; i++) {
             server.create('contributor', { node: nodes[i], users: currentUser, index: 11 });
         }
@@ -53,7 +54,8 @@ module('Acceptance | dashboard', hooks => {
         const currentUser = server.create('user');
         server.create('root', { currentUser });
         const nodes = server.createList('node', 10, {}, 'withContributors');
-        server.loadFixtures('nodes');
+        server.create('node', { id: 'z3sg2', linkedNodeIds: ['1', '2', '3', '4', '5'], title: 'NNW' });
+        server.create('node', { id: '57tnq', linkedNodeIds: ['6', '7', '8', '9', '10'], title: 'Popular' });
         await visit('/dashboard');
         assert.dom('img[alt*="Missing translation"]').doesNotExist();
         for (const node of nodes) {
@@ -95,7 +97,8 @@ module('Acceptance | dashboard', hooks => {
     test('user has many projects', async function(assert) {
         const currentUser = server.create('user');
         const nodes = server.createList('node', 30, {}, 'withContributors');
-        server.loadFixtures('nodes');
+        server.create('node', { id: 'z3sg2', linkedNodeIds: ['1', '2', '3', '4', '5'], title: 'NNW' });
+        server.create('node', { id: '57tnq', linkedNodeIds: ['6', '7', '8', '9', '10'], title: 'Popular' });
         for (const node of nodes) {
             server.create('contributor', { node, users: currentUser, index: 11 });
         }