+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - Update ember-cli-addon-docs dependency
 - Misc:
     - upgrade to ember(-(cli|data))@~3.3.0
+- DX:
+    - No more mirage fixtures
 
 ## [0.7.0] - 2018-08-07
 ### Added